pyproject.toml
# Created by: Mustafa Can Caliskan
# Date: 2026-04-25

[project]
name = "pdgraph-bench"
version = "0.1.0"
description = "Portable benchmark for open-source distributed property-graph databases under mixed workloads and elastic scale-out"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Mustafa Can Çalışkan" }]
license = { text = "MIT" }

dependencies = [
    "anyio>=4.3",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "pyarrow>=15.0",
    "hdrhistogram>=0.10",
    "psutil>=5.9",
    "numpy>=1.26",
    "scipy>=1.12",
    "pandas>=2.2",
    "matplotlib>=3.8",
    "typer>=0.9",
    "structlog>=24.1",
    "aiohttp>=3.9",
    "tabulate>=0.10.0",
]

[project.optional-dependencies]
drivers = [
    "nebula3-python>=3.5",
    "python-arango>=7.9",
    "pydgraph>=21.3",
    "gremlinpython>=3.7",
    "httpx>=0.27",
]
deploy = [
    "fabric>=3.2",
    "invoke>=2.2",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.3",
    "black>=24.3",
    "mypy>=1.9",
]
analysis = [
    "jupyter>=1.0",
    "nbconvert>=7.16",
    "ipykernel>=6.29",
    "tabulate>=0.9",
]

[project.scripts]
pdgraph-bench = "graph_bench.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/graph_bench"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
packages = ["graph_bench"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
markers = [
    "integration: marks integration tests (deselect with '-m \"not integration\"')",
]
